Include all preview browse products in detail tab. Fixes #1216. (#1241)
* Create get_displayed_browse_products to return a list of browse products
urls. These urls will be used to display preview images in detail tab.

* Use flexslider as the carousel for preview images and also update css
styling to make it responsive.

* Add styling to wrap the overflow text of the detail entry values.

* Fixed the test due to newly add class for detail entry values to wrap
the overflow text.

* Update preview image styling to make sure it won't take up the full width
when window is narrow.

* Update styling to keep the transition animation of preview images nav
buttons when window width is between 767-860px.

* Update styling to keep the transition animation of preview images nav
buttons when window width is between 768-860px.

* Set proper spacing between detail tab preview image and the nav buttons.
Set proper spacing between detail tab preview guide message and the nav buttons.

* Include LICENSE.md for flexslider

* Update the styling so that nav buttons of the detail preview images
won't be wrapped into two rows when the screen is narrow.

* Fixed the twitching issue when the cursor is at the border of preview
image and nav buttons.

* Position "<" & ">" buttons out of the preview image. Now they won't overlap
with the preview image.

* When detail preview image moves to the top, position "<" & ">" near the
image instead of the edges of the screen. Shrink the size of "..." when
screen is getting narrower.

* Fine tune the position of "<" & ">" when detail preview image is above
the metadata info when screen is narrow.

* Create adjustDetailImgNavBtns to adjust the positions of detail preview
image nav buttons "<" & ">". Make sure they stay right next to the image
at both sides (even when the window is resizing).
- In detail.js (line 220-234), create adjustDetailImgNavBtns to make sure
  preview image nav buttons stay right next to the image.

* Fixed the inconsistent function name adjustDetailImgNavBtns in detail.js
and opus.js.

* Shrink the detail preview image size and increase the distance between
"<" & ">" and the image.

* Shrink the detail preview image size to 85% so that ">" is a bit further
away from feedback tab.

* Init flexslider after html are loaded in detail tab, that way the arrow
won't show up until it's finally positioned right next to the image.

* Update the stylings for detail tab when screen is short and narrow.

* Update get_displayed_browse_products to handle the case when there is
no preview image for an opus id.
- In file_utils.py (line 247), set selected_browse_products to an empty
  array when the opus id has no preview images at all.

* Update get_displayed_browse_products to return settings.THUMBNAIL_NOT_FOUND
when an observation doesn't have a preview image. This will make sure
preview not found image is properly displayed in detail tab.

* Update styling to make sure when there is only one detail preview image,
the image is as close as to the left container.

* Clean up code style/comments based on Github feedback.

* Use cdn for flexslider.

* Update comments with the reasoning of why we select offset 40 in adjustDetailImgNavBtns.

* Update comments in adjustDetailImgNavBtns (line 224-228)