Fixed multiple tooltips issues (1265, 1273, 1274) (#1308)
* Update the position of browse table row tooltips, instead of sticking with
the cursor, it will now stay at the top border of the row.
- In utils.js (line 144-169), create setTableColumnToolTipPostion to
set the position of the tooltips for browse table column to stay at the
top border of the row.
- In browse.js (line 1911), call setTableColumnToolTipPostion to set
the position of the tooltips for browse table column.

* Add a try catch block to bypass the error message if the table sorting
(redrawing) happened right at the moment when a tooltips instance is
about to reposition. This error message will complain about we try access
an uninitialized element (due to table redrawing). We don't need to worry
about this error because the tooltip instance will get initialized again
once redrawing is done and no weird behavior with this. (line 140-149
in utiles.js)

* Fixed the tooltip positoin of the browse gallery "Detail" icon.
- In browse.js (line 1902-1906), move the tooltip position next to the
"Detail" icon.

* Fixed the tooltip positoin of the browse gallery binoculars icon.
- In browse.js (line 1916-1921), move the tooltip position to the top
of the binoculars icon.

* Fine tune the tooltip position of "Detail" icon so that it shows of at
the upper middle of the icon. (line 1905 in browse.js)

* Update the settings for ".op-metadatabox-edit-tooltip" tooltips, now they
will show up at the left of trash and "+" icons, and won't block the
metadata edit menu.

* Fixed the positions of the tooltips for cart, hamburger, prev, and next
icons in metadata box, now tooltips will show up at the centered top of
the icons.
- In browse.js (line 3012-3021), move "op-metadatabox-tooltip" class to
the icon <i> tags, and apply title on those <i> tags as well. This will
make sure tooltips stay with <i> tags. Previous the class name is added
the parent <a> tags, tooltips position is a bit off when the children <i>
tags have stylings related to the positions.

* Increase the delay of tooltips on the gallery and table view.
- In browse.js (line 1891, 1932), create a constant galleryAndTableTooltipsDelay
to store the longer delay for tooltips of the gallery and table view.

* Fixed the code style of missing semicolon at line 148 of utils.js

* Change "op-browse-gallery-bino-tooltip" to "op-browse-gallery-binocular-tooltip"
in browse.js